KVM: SVM: Zero ASID generation on flush (instead of decrementing)
Zero the ASID generation when using it to induce a TLB "flush" by forcing
a new ASID.  Decrementing the generation is bizarre and unnecessary, as
the per-CPU generation starts counting at '1' and is a 64-bit value, i.e.
the per-CPU generation will never be '0'.

Remove a semi-related comment about "fixing" ASID generation wraparound.
The generation is a 64-bit value; even if KVM managed to trigger a new
ASID generation on every clock tick, that still allows for a few billion
years before wrap will occur.

Signed-off-by: Sean Christopherson <seanjc@google.com>