KVM: x86/mmu: Initialize MMU constants once per vCPU
When possible, initialize kvm_mmu fields that are constants once per vCPU
during vCPU creation.  Now that KVM uses different MMUs for L1 versus L2
when TDP is enabled, the majority of callbacks are constant for a given
MMU.

Nested NPT and IA32 shadow paging have a few constant callbacks, but not
as many due to the need to handle both 32-bit and 64-bit MMUs (and
non-paging for IA32).

Note, nested NPT is always an indirect MMU, but extracting that single
flag out of the common shadow MMU flows is not remotely worthwhile (IA32
shadow paging is direct when the guest has paging disabled).

Avoiding the unnecessary refresh on MMU initialization can shave a few
microseconds from guest boot if the guest is doing silly things like
using SMM as a trampoline.

Opportunistically change kvm_mmu_create() to return '0' unconditionally
in its happy path to make it obvious that it's a happy path.

Reported-by: Hao Peng <flyingpenghao@gmail.com>
Signed-off-by: Sean Christopherson <seanjc@google.com>